Rename common/coverart.js to common/loadArtwork.js
This is used for event art too, and "load" makes it clearer what it's for.
