+++ root/static/scripts/common/loadArtwork.js
